% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history_anno}
\alias{history_anno}
\title{History of Amherst College annotated text}
\format{
A list (also cnlp_annotation class, see cleanNLP package) with 3 objects on the following variables .
\itemize{
\item{\code{token}} {tibble of token information}
\item{\code{entity}}{tibble containing entity information}
\item{\code{document}}{document information}
}
}
\source{
See https://archive.org/details/historyofamherst00tyleiala for the book in a variety of formats.
}
\usage{
data(history_anno)
}
\description{
Annotated entities from the cleanNLP package.
}
\keyword{datasets}
