% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history_figures}
\alias{history_figures}
\title{History of Amherst College figures}
\format{
A data frame with XX observations on the following variables.
\itemize{
\item{\code{pdf_pg}} {The page the figure is on in the PDF}
\item{\code{caption}}{The caption of the figure}
}
}
\source{
See https://archive.org/details/historyofamherst00tyleiala for the book in a variety of formats.
}
\usage{
data(history_figures)
}
\description{
This table details the page number in the PDF of the 20 images present in the book
}
\keyword{datasets}
