% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history_subtitles}
\alias{history_subtitles}
\title{History of Amherst College page sub titles}
\format{
A data frame with XX observations on the following variables.
\itemize{
\item{\code{chapter}} {The chapter the heading is present in}
\item{\code{page_number}}{The page number the heading occurs on}
\item{\code{page_header}}{The actual string value of the heading}
}
}
\source{
See https://archive.org/details/historyofamherst00tyleiala for the book in a variety of formats.
}
\usage{
data(history_subtitles)
}
\description{
This is a table of the headings that accompany the page number on each page.
}
\keyword{datasets}
