% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history_tables}
\alias{history_tables}
\title{History of Amherst College tables}
\format{
A data frame with 15 observations on the following variables.
\itemize{
\item{\code{pdf_pg}} {The page the table starts on in the PDF}
\item{\code{book_pg}}{The page of the actual book the table starts on}
\item{\code{title}}{The name/description of the table}
}
}
\source{
See https://archive.org/details/historyofamherst00tyleiala for the book in a variety of formats.
}
\usage{
data(history_tables)
}
\description{
This table details the page number in the book and the PDF of the 15 tables present in the book
}
\keyword{datasets}
