% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history_toc}
\alias{history_toc}
\title{History of Amherst College Table of Contents}
\format{
A data frame with 29 observations on the following variables.
\itemize{
\item{\code{title}} {The title of the chapter}
\item{\code{page_number}}{The page number where the chapter starts}
\item{\code{chapter}}{The chapter number}
}
}
\source{
See https://archive.org/details/historyofamherst00tyleiala for the book in a variety of formats.
}
\usage{
data(history_toc)
}
\description{
This table contains the page number and the titles for every chapter in the book including the Appendix
}
\keyword{datasets}
