% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history_text}
\alias{history_text}
\title{History of Amherst College text}
\format{
A data frame with XX observations on the following variables.
\itemize{
\item{\code{file_name}} {File name of scanned image (in extdata)}
\item{\code{house_or_not}}{House string in description (coarse way to identify houses)}
}
}
\source{
See https://archive.org/details/historyofamherst00tyleiala for the book in a variety of formats.
}
\usage{
data(history_text)
}
\description{
XX More description?
}
\keyword{datasets}
