% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{history_text}
\alias{history_text}
\title{History of Amherst College text}
\format{
A data frame with XX observations on the following variables.
\itemize{
\item{\code{text}} {text of the book}
\item{\code{chapter}{chapter number}
\item{\code{paragraph}{paragraph in the chapter}
\item{\code{page_number}{page number in the book}
}
}
\source{
See https://archive.org/details/historyofamherst00tyleiala for the book in a variety of formats.
}
\usage{
data(history_text)
}
\description{
The text of the 1873 book "A History of Amherst College" by William Seymour Tyler.
}
\keyword{datasets}
